+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Released on FIXME: ??
 - Bugfix:
   - Fixed wrong text wrap in log box
   - Fixed error message not being shown after an upload failure
+  - Fixed default protocol not being loaded from config
   - [Issue 23](https://github.com/veeso/termscp/issues/23): Remove created file if transfer failed or was abrupted
 - Dependencies:
   - Added `tui-realm 0.2.2`